% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_sigma.R
\name{optimal_sigma}
\alias{optimal_sigma}
\title{Compute the optimal sigma for the volume of the intersection}
\usage{
optimal_sigma(pointsA, pointsB, epsilon = 0.01, n = 10000, n_reps = 5)
}
\arguments{
\item{pointsA}{A matrix of points}

\item{pointsB}{A matrix of points}

\item{epsilon}{The error between the estimated volume and the PAVO volume
to determine the optimal sigma}

\item{n}{The number of samples to use}

\item{n_reps}{The number of repetitions to use}
}
\value{
The optimal sigma for the volume of the intersection
}
\description{
Compute the optimal sigma for the volume of the intersection
}
