% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voloverlap.R
\name{volume_overlap}
\alias{volume_overlap}
\title{Volume overlap estimation for two sets of points}
\usage{
volume_overlap(points_a, points_b, sigma = NULL, n_sigma = 5)
}
\arguments{
\item{points_a}{A matrix of points}

\item{points_b}{A matrix of points}

\item{sigma}{The standard deviation of the Gaussian kernel. If NULL, the
volume overlap is estimated for a range of sigmas.}

\item{n_sigma}{The number of sigmas to use for estimating the volume overlap.
Only used if sigma is NULL.}
}
\value{
A data frame with the volume overlap estimation
}
\description{
Estimate the volume overlap between two sets of points
}
