% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voloverlap.R
\name{volume_overlap}
\alias{volume_overlap}
\title{Volume overlap estimation for two sets of points}
\usage{
volume_overlap(pointsA, pointsB, sigma = 1, n = 10000, n_reps = 5)
}
\arguments{
\item{pointsA}{A matrix of points}

\item{pointsB}{A matrix of points}

\item{sigma}{The standard deviation of the Gaussian kernel}

\item{n}{The number of samples to use}

\item{n_reps}{The number of repetitions to use}
}
\value{
A data frame with the volume overlap estimation
}
\description{
Estimate the volume overlap between two sets of points
}
